Update test_fitness.py
We are trying out a new fitness score based on the Normalized Euclidean Distance instead of a Chi-Squared calculation as we have begun to question the validity of that measure as a test of fit given that they are not intended for discrete points. For instance, a run with a large number of genes does not behave the same as one with few genes. So instead we are treating this as a distance metric between two points in an N- dimensional Euclidean space to account for the number of genes.